+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-### Fixed
+### Added
 
-- Publish edition-aware OpenAPI and compatibility contracts ([monica-openapi-edition-contracts](https://github.com/unbraind/monica-cli/blob/master/.agents/pm/issues/monica-openapi-edition-contracts.toon))
+- Publish edition-aware OpenAPI and compatibility contracts ([monica-openapi-edition-contracts](https://github.com/unbraind/monica-cli/blob/master/.agents/pm/features/monica-openapi-edition-contracts.toon))
 
 ## 2026.7.26 - 2026-07-26